statistics and calacultors
