+++ src/handlers/calculator-handler/oi3-real.js
@@ -1,14 +1,12 @@
-const moment = require('moment');
+﻿const moment = require('moment');
 require('twix');
-const _ = require('underscore');
 const mongoose = require('mongoose');
-const { ObjectID } = require('mongodb');
 const CameraLocationModel = require('../../models/data/camera-location-model');
 const CameraLocationState = require('../../models/const/camera-location-state');
 const AnnotationModel = require('../../models/data/annotation-model');
 const AnnotationState = require('../../models/const/annotation-state');
+const ProjectTrip = require('../../models/data/project-trip-model');
 const SpeciesModel = require('../../models/data/species-model');
-const DataFieldModel = require('../../models/data/data-field-model');
 
 const fetchCameraLocations = async cameraLocationIds => {
   const cameraLocations = await CameraLocationModel.where({
@@ -102,76 +100,13 @@ module.exports = async (req, res) => {
   const annotationQuery = getAnnotationQuery(form);
   const annotations = await annotationQuery;
 
-  const organismIdDataField = await DataFieldModel.findOne({
-    'title.zh-TW': '個體 ID',
+  const trips = await ProjectTrip.find({
+    'studyAreas.cameraLocations.cameraLocation': {
+      $in: cameraLocationIds,
+    },
   });
 
-  const userid = [];
-  cameraLocationIds.forEach(stringId => {
-    userid.push(new ObjectID(stringId));
-  });
-
-  const timesTT = await AnnotationModel.aggregate([
-    {
-      $match: { cameraLocation: { $in: userid } },
-    },
-
-    {
-      $sort: {
-        time: -1.0,
-      },
-    },
-    {
-      $group: {
-        _id: '$cameraLocation',
-        starttimes: {
-          $first: {
-            $arrayElemAt: ['$rawData', 4.0],
-          },
-        },
-        endtimes: {
-          $last: {
-            $arrayElemAt: ['$rawData', 4.0],
-          },
-        },
-      },
-    },
-  ]);
-  // console.log(timesTT)
-  const timesArray = [];
-  const totalTest = {};
-  await timesTT.forEach(t => {
-    const Start = moment(t.starttimes).format('YYYY-MM-DDTHH:mm:ss');
-    const End = moment(t.endtimes).format('YYYY-MM-DDTHH:mm:ss');
-
-    if (moment(End).isAfter(Start)) {
-      const durationsT = moment(End).diff(Start);
-      totalTest[t._id] = durationsT;
-
-      // console.log(totalTest)
-
-      timesArray.push({
-        id: t._id,
-        start: Start,
-        end: End,
-      });
-    } else {
-      const durationsT = moment(Start).diff(End);
-      totalTest[t._id] = durationsT;
-
-      // console.log(totalTest)
-
-      timesArray.push({
-        id: t._id,
-        start: End,
-        end: Start,
-      });
-    }
-  });
-
-  const result = _.groupBy(timesArray, 'id');
-
-  /* const times = {};
+  const times = {};
   const totalTime = {};
   trips.forEach(t => {
     t.studyAreas.forEach(s => {
@@ -196,15 +131,14 @@ module.exports = async (req, res) => {
         },
       );
     });
-  }); */
+  });
 
   const data = [];
-  const organismIdList = [];
   if (range === 'month') {
     const monthList = getMonthRange(startDateTime, endDateTime);
     species.forEach(({ _id: s }) => {
       cameraLocations.forEach(c => {
-        const workingCameraRange = result[c._id];
+        const workingCameraRange = times[c.name];
         monthList.forEach(m => {
           let totalPics = 0;
           let lastValidAnnotationTime;
@@ -215,44 +149,29 @@ module.exports = async (req, res) => {
               ({ species: { _id: annotationSpeicesId } }) =>
                 `${annotationSpeicesId}` === `${s}`,
             )
-            .forEach(({ time, fields }) => {
-              // 考慮個體ID
-              let countOrganism = false;
-              if (fields.length > 0) {
-                const hasOrganismId = fields.find(
-                  x =>
-                    x.dataField.toString() ===
-                    organismIdDataField._id.toString(),
-                );
-                const organismId = hasOrganismId.value.text;
-                if (organismIdList.indexOf(organismId) === -1) {
-                  organismIdList.push(organismId);
-                  countOrganism = true;
-                }
+            .forEach(({ time }) => {
+              if (!lastValidAnnotationTime) {
+                lastValidAnnotationTime = time;
+                totalPics = 1;
               }
-              if (countOrganism === true) {
-                if (!lastValidAnnotationTime) {
-                  lastValidAnnotationTime = time;
-                  totalPics = 1;
-                }
-                if (
-                  moment(time).diff(lastValidAnnotationTime) >
-                  calculateTimeIntervel
-                ) {
-                  lastValidAnnotationTime = time;
-                  totalPics += 1;
-                }
+
+              if (
+                moment(time).diff(lastValidAnnotationTime) >
+                calculateTimeIntervel
+              ) {
+                lastValidAnnotationTime = time;
+                totalPics += 1;
               }
             });
 
           // 整理時數
-          const beginT = moment(m).startOf('month');
-          const endT = moment(m).endOf('month');
+          const begin = moment(m).startOf('month');
+          const end = moment(m).endOf('month');
           let hours = 0;
 
-          workingCameraRange.forEach(({ start, end }) => {
-            const range1 = moment(start).twix(end);
-            const range2 = moment(beginT).twix(endT);
+          workingCameraRange.forEach(({ startTime, endTime }) => {
+            const range1 = moment(startTime).twix(endTime);
+            const range2 = moment(begin).twix(end);
 
             const rr = range1.intersection(range2);
             const duration = rr._end.diff(rr._start);
@@ -262,6 +181,7 @@ module.exports = async (req, res) => {
           });
 
           const totalHours = parseFloat(hours.toFixed(4));
+
           data.push({
             species: s,
             cameraLocationId: c._id,
